+++ exporters/jaeger/internal/gen-go/jaeger/jaeger.go
@@ -8,8 +8,9 @@ import (
 	"database/sql/driver"
 	"errors"
 	"fmt"
-	"go.opentelemetry.io/otel/exporters/trace/jaeger/internal/third_party/thrift/lib/go/thrift"
 	"time"
+
+	"go.opentelemetry.io/otel/exporters/jaeger/internal/third_party/thrift/lib/go/thrift"
 )
 
 // (needed to ensure safety because of naive import list construction.)